addressed pull request fixes plus cleaned up model.py
